refactor: remove legacy builder/client/http stack
