--- src/ArduinoIoTCloud.ts
@@ -1,15 +1,0 @@
-import { CloudOptions } from './CloudOptions';
-import { ICloudClientBuilder } from './builder/ICloudClientBuilder';
-import { IArduinoIoTCloudFactory } from './builder/IArduinoIoTCloudFactory';
-
-export function ArduinoIoTCloudFactory(builders: ICloudClientBuilder[]): IArduinoIoTCloudFactory {
-  return {
-    connect: async (options: any) => {
-      options = { ...CloudOptions.DEFAULT, ...options };
-      const builder = builders.find((b) => b.canBuild(options));
-      if (!builder) throw new Error('connection failed: options not valid');
-
-      return builder.build(options) as any;
-    },
-  };
-}
