--- src/http/IHttpClient.ts
@@ -1,3 +1,0 @@
-export interface IHttpClient {
-  post<T, P = any>(uri: string, body: P, headers?: { [key: string]: string }): Promise<T>;
-}
